+++ tests/rules/test_pacman.py
@@ -29,10 +29,10 @@ def test_match(command):
 @pytest.mark.parametrize('command, return_value', [
     (Command('vim', 'vim: command not found'), PKGFILE_OUTPUT_VIM),
     (Command('sudo vim', 'sudo: vim: command not found'), PKGFILE_OUTPUT_VIM)])
-@patch('thebleep.specific.archlinux.subprocess')
+@patch('thebleep.specific.archlinux.utils.tool_lines')
 @patch.multiple(pacman, create=True, pacman=pacman_cmd)
-def test_match_mocked(subp_mock, command, return_value):
-    subp_mock.check_output.return_value = return_value
+def test_match_mocked(tool_lines, command, return_value):
+    tool_lines.return_value = return_value.splitlines()
     assert match(command)
 
 
@@ -75,8 +75,9 @@ def test_get_new_command(command, new_command, mocker):
     (Command('convert', ''), ['{} -S extra/imagemagick && convert'.format(pacman_cmd)], PKGFILE_OUTPUT_CONVERT),
     (Command('sudo', ''), ['{} -S core/sudo && sudo'.format(pacman_cmd)], PKGFILE_OUTPUT_SUDO),
     (Command('sudo convert', ''), ['{} -S extra/imagemagick && sudo convert'.format(pacman_cmd)], PKGFILE_OUTPUT_CONVERT)])
-@patch('thebleep.specific.archlinux.subprocess')
+@patch('thebleep.specific.archlinux.utils.tool_lines')
 @patch.multiple(pacman, create=True, pacman=pacman_cmd)
-def test_get_new_command_mocked(subp_mock, command, new_command, return_value):
-    subp_mock.check_output.return_value = return_value
+def test_get_new_command_mocked(tool_lines, command, new_command,
+                                return_value):
+    tool_lines.return_value = return_value.splitlines()
     assert get_new_command(command) == new_command