+++ tests/rules/test_pacman_not_found.py
@@ -26,9 +26,9 @@ def test_match(command):
     Command('yaourt -S llc', 'error: target not found: llc'),
     Command('pacman llc', 'error: target not found: llc'),
     Command('sudo pacman llc', 'error: target not found: llc')])
-@patch('thebleep.specific.archlinux.subprocess')
-def test_match_mocked(subp_mock, command):
-    subp_mock.check_output.return_value = PKGFILE_OUTPUT_LLC
+@patch('thebleep.specific.archlinux.utils.tool_lines')
+def test_match_mocked(tool_lines, command):
+    tool_lines.return_value = PKGFILE_OUTPUT_LLC.splitlines()
     assert match(command)
 
 
@@ -50,7 +50,7 @@ def test_get_new_command(command, fixed):
     (Command('yaourt -S llc', 'error: target not found: llc'), ['yaourt -S extra/llvm', 'yaourt -S extra/llvm35']),
     (Command('pacman -S llc', 'error: target not found: llc'), ['pacman -S extra/llvm', 'pacman -S extra/llvm35']),
     (Command('sudo pacman -S llc', 'error: target not found: llc'), ['sudo pacman -S extra/llvm', 'sudo pacman -S extra/llvm35'])])
-@patch('thebleep.specific.archlinux.subprocess')
-def test_get_new_command_mocked(subp_mock, command, fixed):
-    subp_mock.check_output.return_value = PKGFILE_OUTPUT_LLC
+@patch('thebleep.specific.archlinux.utils.tool_lines')
+def test_get_new_command_mocked(tool_lines, command, fixed):
+    tool_lines.return_value = PKGFILE_OUTPUT_LLC.splitlines()
     assert get_new_command(command) == fixed